rename to keep same naming-convention as for readungridded
